cchart.c: Don't deduplicate values
This fixes charts for data sources that might give back the same
value repeatably. Most obvious to see when trying to monitor the
disk or swap (the values may not change for a long time).

This still does not allow for true static value assignments.

Testcase:
```
layout ":center" {
	chart {
		style = "disk"
		interval = 2000
		value = (Disk("/", "%used") / 100)
	}
}

#CSS
window#sfwbar {
	-GtkWidget-direction: top;
}

chart {
	color: yellow;
	background-color: black;
	min-width: 50px;
	min-height: 28px;
	margin: 0 1px;
	border: 1px inset rgba(128, 128, 128, 0.5);
}

chart#disk {
	color: purple;
}
```