bar: do not use quoted "layout" as CSS name
I'm not sure why this string included escaped quotes. The change was made in 34c9d917da883f64a63afbd2cbdae53b0fa09a69, but doesn't seem to make sense to me--the only thing this string is used for, from what I can tell, is the CSS name via `gtk_widget_set_name`. And including quotes in such a name makes it impossible to refer to from CSS: `#layout` does not match it, and `#"layout"` is not valid syntax.

This change allows `#layout` to be referenced in CSS.